requirements.txt
influxdb-client
speedtest-cli
requests
configparser